refactor(tests): Reorganize local tests to better match reality
