+++ test/local/crypto/base32_tests.js
@@ -3,7 +3,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 const assert = require('insist')
-const base32 = require('../../lib/crypto/base32')
+const base32 = require('../../../lib/crypto/base32')
 
 describe('base32', () => {
   it('takes 1 integer argument, returns a function', () => {
