+++ test/local/crypto/butil_tests.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const assert = require('assert')
-const butil = require('../../lib/crypto/butil')
+const butil = require('../../../lib/crypto/butil')
 
 describe('butil', () => {
