+++ test/local/routes/account_routes.js
@@ -7,19 +7,18 @@
 var sinon = require('sinon')
 
 const assert = require('insist')
-var mocks = require('../mocks')
-var getRoute = require('../routes_helpers').getRoute
+var mocks = require('../../mocks')
+var getRoute = require('../../routes_helpers').getRoute
 var proxyquire = require('proxyquire')
 
-var P = require('../../lib/promise')
+var P = require('../../../lib/promise')
 var uuid = require('uuid')
 var crypto = require('crypto')
 var isA = require('joi')
-var error = require('../../lib/error')
-var log = require('../../lib/log')
+var error = require('../../../lib/error')
+var log = require('../../../lib/log')
 
 var TEST_EMAIL = 'foo@gmail.com'
-var TEST_EMAIL_INVALID = 'example@dotless-domain'
 
 var makeRoutes = function (options, requireMocks) {
   options = options || {}
@@ -41,17 +40,17 @@ var makeRoutes = function (options, requireMocks) {
   config.signinUnblock = config.signinUnblock || {}
 
   var log = options.log || mocks.mockLog()
-  var Password = options.Password || require('../../lib/crypto/password')(log, config)
+  var Password = options.Password || require('../../../lib/crypto/password')(log, config)
   var db = options.db || mocks.mockDB()
-  var isPreVerified = require('../../lib/preverifier')(error, config)
+  var isPreVerified = require('../../../lib/preverifier')(error, config)
   var customs = options.customs || {
     check: function () { return P.resolve(true) }
   }
-  var checkPassword = options.checkPassword || require('../../lib/routes/utils/password_check')(log, config, Password, customs, db)
-  var push = options.push || require('../../lib/push')(log, db, {})
-  return proxyquire('../../lib/routes/account', requireMocks || {})(
+  var checkPassword = options.checkPassword || require('../../../lib/routes/utils/password_check')(log, config, Password, customs, db)
+  var push = options.push || require('../../../lib/push')(log, db, {})
+  return proxyquire('../../../lib/routes/account', requireMocks || {})(
     log,
-    require('../../lib/crypto/random'),
+    require('../../../lib/crypto/random'),
     P,
     uuid,
     isA,
@@ -64,7 +63,7 @@ var makeRoutes = function (options, requireMocks) {
     isPreVerified,
     checkPassword,
     push,
-    options.devices || require('../../lib/devices')(log, db, push)
+    options.devices || require('../../../lib/devices')(log, db, push)
   )
 }
 
@@ -81,236 +80,6 @@ function runTest (route, request, assertions) {
   })
   .then(assertions)
 }
-
-describe('/recovery_email/status', function () {
-  var config = {}
-  var mockDB = mocks.mockDB()
-  var pushCalled
-  var mockLog = mocks.mockLog({
-    increment: function (name) {
-      if (name === 'recovery_email_reason.push') {
-        pushCalled = true
-      }
-    }
-  })
-  var accountRoutes = makeRoutes({
-    config: config,
-    db: mockDB,
-    log: mockLog
-  })
-  var route = getRoute(accountRoutes, '/recovery_email/status')
-
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: uuid.v4('binary').toString('hex'),
-      email: TEST_EMAIL
-    }
-  })
-
-  describe('invalid email', function () {
-    var mockRequest = mocks.mockRequest({
-      credentials: {
-        email: TEST_EMAIL_INVALID
-      }
-    })
-
-    it('unverified account', function () {
-      mockRequest.auth.credentials.emailVerified = false
-
-      return runTest(route, mockRequest).then(() => assert.ok(false), function (response) {
-        assert.equal(mockDB.deleteAccount.callCount, 1)
-        assert.equal(mockDB.deleteAccount.firstCall.args[0].email, TEST_EMAIL_INVALID)
-        assert.equal(response.errno, error.ERRNO.INVALID_TOKEN)
-      })
-      .then(function () {
-        mockDB.deleteAccount.reset()
-      })
-    })
-
-    it('verified account', function () {
-      mockRequest.auth.credentials.uid = uuid.v4('binary').toString('hex')
-      mockRequest.auth.credentials.emailVerified = true
-      mockRequest.auth.credentials.tokenVerified = true
-
-      return runTest(route, mockRequest, function (response) {
-        assert.equal(mockDB.deleteAccount.callCount, 0)
-        assert.deepEqual(response, {
-          email: TEST_EMAIL_INVALID,
-          verified: true,
-          emailVerified: true,
-          sessionVerified: true
-        })
-      })
-    })
-  })
-
-
-  it('valid email, verified account', function () {
-    pushCalled = false
-    var mockRequest = mocks.mockRequest({
-      credentials: {
-        uid: uuid.v4('binary').toString('hex'),
-        email: TEST_EMAIL,
-        emailVerified: true,
-        tokenVerified: true
-      },
-      query: {
-        reason: 'push'
-      }
-    })
-
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(pushCalled, true)
-
-      assert.deepEqual(response, {
-        email: TEST_EMAIL,
-        verified: true,
-        emailVerified: true,
-        sessionVerified: true
-      })
-    })
-  })
-
-  it('verified account, verified session', function () {
-    mockRequest.auth.credentials.emailVerified = true
-    mockRequest.auth.credentials.tokenVerified = true
-
-    return runTest(route, mockRequest, function (response) {
-      assert.deepEqual(response, {
-        email: TEST_EMAIL,
-        verified: true,
-        sessionVerified: true,
-        emailVerified: true
-      })
-    })
-  })
-
-  it('verified account, unverified session, must verify session', function () {
-    mockRequest.auth.credentials.emailVerified = true
-    mockRequest.auth.credentials.tokenVerified = false
-    mockRequest.auth.credentials.mustVerify = true
-
-    return runTest(route, mockRequest, function (response) {
-      assert.deepEqual(response, {
-        email: TEST_EMAIL,
-        verified: false,
-        sessionVerified: false,
-        emailVerified: true
-      })
-    })
-  })
-
-  it('verified account, unverified session, neednt verify session', function () {
-    mockRequest.auth.credentials.emailVerified = true
-    mockRequest.auth.credentials.tokenVerified = false
-    mockRequest.auth.credentials.mustVerify = false
-
-    return runTest(route, mockRequest, function (response) {
-      assert.deepEqual(response, {
-        email: TEST_EMAIL,
-        verified: true,
-        sessionVerified: false,
-        emailVerified: true
-      })
-    })
-  })
-})
-
-describe('/recovery_email/resend_code', () => {
-  const config = {}
-  const mockDB = mocks.mockDB()
-  const mockLog = mocks.mockLog()
-  mockLog.flowEvent = sinon.spy(() => {
-    return P.resolve()
-  })
-  const mockMailer = mocks.mockMailer()
-  const mockMetricsContext = mocks.mockMetricsContext()
-  const accountRoutes = makeRoutes({
-    config: config,
-    db: mockDB,
-    log: mockLog,
-    mailer: mockMailer
-  })
-  const route = getRoute(accountRoutes, '/recovery_email/resend_code')
-
-  it('verification', () => {
-    const mockRequest = mocks.mockRequest({
-      log: mockLog,
-      metricsContext: mockMetricsContext,
-      credentials: {
-        uid: uuid.v4('binary').toString('hex'),
-        email: TEST_EMAIL,
-        emailVerified: false,
-        tokenVerified: false,
-        uaBrowser: 'Firefox',
-        uaBrowserVersion: 52,
-        uaOS: 'Mac OS X',
-        uaOSVersion: '10.10'
-      },
-      query: {},
-      payload: {
-        metricsContext: {
-          flowBeginTime: Date.now(),
-          flowId: 'F1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF103'
-        }
-      }
-    })
-    mockLog.flowEvent.reset()
-
-    return runTest(route, mockRequest, response => {
-      assert.equal(mockLog.flowEvent.callCount, 1, 'log.flowEvent called once')
-      assert.equal(mockLog.flowEvent.args[0][0].event, 'email.verification.resent')
-
-      assert.equal(mockMailer.sendVerifyCode.callCount, 1)
-      const args = mockMailer.sendVerifyCode.args[0]
-      assert.equal(args[2].uaBrowser, 'Firefox')
-      assert.equal(args[2].uaBrowserVersion, '52')
-      assert.equal(args[2].uaOS, 'Mac OS X')
-      assert.equal(args[2].uaOSVersion, '10.10')
-      assert.strictEqual(args[2].uaDeviceType, undefined)
-    })
-  })
-
-  it('confirmation', () => {
-    const mockRequest = mocks.mockRequest({
-      log: mockLog,
-      metricsContext: mockMetricsContext,
-      credentials: {
-        uid: uuid.v4('binary').toString('hex'),
-        email: TEST_EMAIL,
-        emailVerified: true,
-        tokenVerified: false,
-        uaBrowser: 'Firefox',
-        uaBrowserVersion: '50',
-        uaOS: 'Android',
-        uaOSVersion: '6',
-        uaDeviceType: 'tablet'
-      },
-      query: {},
-      payload: {
-        metricsContext: {
-          flowBeginTime: Date.now(),
-          flowId: 'F1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF103'
-        }
-      }
-    })
-    mockLog.flowEvent.reset()
-
-    return runTest(route, mockRequest, response => {
-      assert.equal(mockLog.flowEvent.callCount, 1, 'log.flowEvent called once')
-      assert.equal(mockLog.flowEvent.args[0][0].event, 'email.confirmation.resent')
-
-      assert.equal(mockMailer.sendVerifyLoginEmail.callCount, 1)
-      const args = mockMailer.sendVerifyLoginEmail.args[0]
-      assert.equal(args[2].uaBrowser, 'Firefox')
-      assert.equal(args[2].uaBrowserVersion, '50')
-      assert.equal(args[2].uaOS, 'Android')
-      assert.equal(args[2].uaOSVersion, '6')
-      assert.strictEqual(args[2].uaDeviceType, 'tablet')
-    })
-  })
-
-})
 
 describe('/account/reset', function () {
   it('should do things', () => {
@@ -406,320 +175,6 @@ describe('/account/reset', function () {
       assert.deepEqual(args[0].tokenId, keyFetchTokenId, 'argument was key fetch token')
       assert.deepEqual(args[0].uid, uid, 'keyFetchToken.uid was correct')
       assert.equal(mockMetricsContext.stash.thisValues[1], mockRequest, 'this was request')
-    })
-  })
-})
-
-describe('/account/device', function () {
-  var config = {}
-  var uid = uuid.v4('binary')
-  var deviceId = crypto.randomBytes(16)
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      deviceCallbackPublicKey: '',
-      deviceCallbackURL: '',
-      deviceId: deviceId,
-      deviceName: 'my awesome device',
-      deviceType: 'desktop',
-      tokenId: crypto.randomBytes(16),
-      uid: uid
-    },
-    payload: {
-      id: deviceId.toString('hex'),
-      name: 'my awesome device'
-    }
-  })
-  var mockDevices = mocks.mockDevices()
-  var mockLog = mocks.spyLog()
-  var accountRoutes = makeRoutes({
-    config: config,
-    devices: mockDevices,
-    log: mockLog
-  })
-  var route = getRoute(accountRoutes, '/account/device')
-
-  it('identical data', function () {
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(mockLog.increment.callCount, 1, 'a counter was incremented')
-      assert.equal(mockLog.increment.firstCall.args[0], 'device.update.spurious')
-
-      assert.deepEqual(response, mockRequest.payload)
-    })
-    .then(function () {
-      mockLog.increment.reset()
-    })
-  })
-
-  it('different data', function () {
-    mockRequest.auth.credentials.deviceId = crypto.randomBytes(16)
-    var payload = mockRequest.payload
-    payload.name = 'my even awesomer device'
-    payload.type = 'phone'
-    payload.pushCallback = 'https://push.services.mozilla.com/123456'
-    payload.pushPublicKey = 'SomeEncodedBinaryStuffThatDoesntGetValidedByThisTest'
-
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(mockLog.increment.callCount, 5, 'the counters were incremented')
-      assert.equal(mockLog.increment.getCall(0).args[0], 'device.update.sessionToken')
-      assert.equal(mockLog.increment.getCall(1).args[0], 'device.update.name')
-      assert.equal(mockLog.increment.getCall(2).args[0], 'device.update.type')
-      assert.equal(mockLog.increment.getCall(3).args[0], 'device.update.pushCallback')
-      assert.equal(mockLog.increment.getCall(4).args[0], 'device.update.pushPublicKey')
-
-      assert.equal(mockDevices.upsert.callCount, 1, 'devices.upsert was called once')
-      var args = mockDevices.upsert.args[0]
-      assert.equal(args.length, 3, 'devices.upsert was passed three arguments')
-      assert.equal(args[0], mockRequest, 'first argument was request object')
-      assert.deepEqual(args[1].tokenId, mockRequest.auth.credentials.tokenId, 'second argument was session token')
-      assert.deepEqual(args[1].uid, uid, 'sessionToken.uid was correct')
-      assert.deepEqual(args[2], mockRequest.payload, 'third argument was payload')
-    })
-    .then(function () {
-      mockLog.increment.reset()
-      mockDevices.upsert.reset()
-    })
-  })
-
-  it('with no id in payload', function () {
-    mockRequest.payload.id = undefined
-
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(mockLog.increment.callCount, 0, 'log.increment was not called')
-
-      assert.equal(mockDevices.upsert.callCount, 1, 'devices.upsert was called once')
-      var args = mockDevices.upsert.args[0]
-      assert.equal(args[2].id, mockRequest.auth.credentials.deviceId.toString('hex'), 'payload.id defaulted to credentials.deviceId')
-    })
-    .then(function () {
-      mockLog.increment.reset()
-      mockDevices.upsert.reset()
-    })
-  })
-
-  it('device updates disabled', function () {
-    config.deviceUpdatesEnabled = false
-
-    return runTest(route, mockRequest, function () {
-      assert(false, 'should have thrown')
-    })
-    .then(() => assert.ok(false), function (err) {
-      assert.equal(err.output.statusCode, 503, 'correct status code is returned')
-      assert.equal(err.errno, error.ERRNO.FEATURE_NOT_ENABLED, 'correct errno is returned')
-    })
-  })
-})
-
-describe('/account/devices/notify', function () {
-  var config = {}
-  var uid = uuid.v4('binary')
-  var mockRequest = mocks.mockRequest({
-    credentials: {
-      uid: uid.toString('hex')
-    }
-  })
-  var pushPayload = {
-    isValid: true,
-    version: 1,
-    command: 'sync:collection_changed',
-    data: {
-      collections: ['clients']
-    }
-  }
-  var mockPush = mocks.mockPush()
-  var validate = sinon.spy(function (payload) { return payload.isValid })
-  var mockAjv = function () {
-    return {
-      compile: function () {
-        return validate
-      }
-    }
-  }
-  var sandbox = sinon.sandbox.create()
-  var mockCustoms = mocks.mockCustoms()
-  var accountRoutes = makeRoutes({
-    config: config,
-    customs: mockCustoms,
-    push: mockPush
-  }, {
-    ajv: mockAjv
-  })
-  var route = getRoute(accountRoutes, '/account/devices/notify')
-
-  it('bad payload', function () {
-    mockRequest.payload = {
-      to: ['bogusid1'],
-      payload: {
-        isValid: false
-      }
-    }
-    return runTest(route, mockRequest, function () {
-      assert(false, 'should have thrown')
-    })
-    .then(() => assert(false), function (err) {
-      assert.equal(validate.callCount, 1, 'ajv validator function was called')
-      assert.equal(mockPush.pushToDevices.callCount, 0, 'mockPush.pushToDevices was not called')
-      assert.equal(err.errno, 107, 'Correct errno for invalid push payload')
-    })
-  })
-
-  it('all devices', function () {
-    mockRequest.payload = {
-      to: 'all',
-      excluded: ['bogusid'],
-      TTL: 60,
-      payload: pushPayload
-    }
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      assert.equal(mockPush.pushToAllDevices.callCount, 1, 'mockPush.pushToAllDevices was called once')
-      var args = mockPush.pushToAllDevices.args[0]
-      assert.equal(args.length, 3, 'mockPush.pushToAllDevices was passed three arguments')
-      assert.equal(args[0], uid.toString('hex'), 'first argument was the device uid')
-      assert.equal(args[1], 'devicesNotify', 'second argument was the devicesNotify reason')
-      assert.deepEqual(args[2], {
-        data: new Buffer(JSON.stringify(pushPayload)),
-        excludedDeviceIds: ['bogusid'],
-        TTL: 60
-      }, 'third argument was the push options')
-    })
-  })
-
-  it('specific devices', function () {
-    mockCustoms.checkAuthenticated.reset()
-    mockRequest.payload = {
-      to: ['bogusid1', 'bogusid2'],
-      TTL: 60,
-      payload: pushPayload
-    }
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      assert.equal(mockPush.pushToDevices.callCount, 1, 'mockPush.pushToDevices was called once')
-      var args = mockPush.pushToDevices.args[0]
-      assert.equal(args.length, 4, 'mockPush.pushToDevices was passed four arguments')
-      assert.equal(args[0], uid.toString('hex'), 'first argument was the device uid')
-      assert.deepEqual(args[1], ['bogusid1', 'bogusid2'], 'second argument was the list of device ids')
-      assert.equal(args[2], 'devicesNotify', 'third argument was the devicesNotify reason')
-      assert.deepEqual(args[3], {
-        data: new Buffer(JSON.stringify(pushPayload)),
-        TTL: 60
-      }, 'fourth argument was the push options')
-    })
-  })
-
-  it('device driven notifications disabled', function () {
-    config.deviceNotificationsEnabled = false
-    mockRequest.payload = {
-      to: 'all',
-      excluded: ['bogusid'],
-      TTL: 60,
-      payload: pushPayload
-    }
-
-    return runTest(route, mockRequest, function () {
-      assert(false, 'should have thrown')
-    })
-    .then(() => assert.ok(false), function (err) {
-      assert.equal(err.output.statusCode, 503, 'correct status code is returned')
-      assert.equal(err.errno, error.ERRNO.FEATURE_NOT_ENABLED, 'correct errno is returned')
-    })
-  })
-
-  it('throws error if customs blocked the request', function () {
-    config.deviceNotificationsEnabled = true
-
-    mockCustoms = mocks.mockCustoms({
-      checkAuthenticated: sandbox.spy(function () {
-        throw error.tooManyRequests(1)
-      })
-    })
-    route = getRoute(makeRoutes({customs: mockCustoms}), '/account/devices/notify')
-
-    return runTest(route, mockRequest, function (response) {
-      assert(false, 'should have thrown')
-    })
-    .then(() => assert(false), function (err) {
-      assert.equal(mockCustoms.checkAuthenticated.callCount, 1, 'mockCustoms.checkAuthenticated was called once')
-      assert.equal(err.message, 'Client has sent too many requests')
-    })
-  })
-
-  it('logs error if no devices found', () => {
-    mockRequest.payload = {
-      to: ['bogusid1', 'bogusid2'],
-      TTL: 60,
-      payload: pushPayload
-    }
-
-    var mockLog = mocks.spyLog()
-    var mockPush = mocks.mockPush({
-      pushToDevices: () => P.reject('Devices ids not found in devices')
-    })
-    var mockCustoms = {
-      checkAuthenticated: () => P.resolve()
-    }
-
-    route = getRoute(makeRoutes({
-      customs: mockCustoms,
-      log: mockLog,
-      push: mockPush
-    }), '/account/devices/notify')
-
-    return runTest(route, mockRequest, function (response) {
-      assert.equal(JSON.stringify(response), '{}', 'response should not throw push errors')
-    })
-  })
-})
-
-describe('/account/device/destroy', function () {
-  it('should work', () => {
-    var uid = uuid.v4('binary')
-    var deviceId = crypto.randomBytes(16).toString('hex')
-    var mockLog = mocks.spyLog()
-    var mockDB = mocks.mockDB()
-    var mockRequest = mocks.mockRequest({
-      credentials: {
-        uid: uid.toString('hex'),
-      },
-      log: mockLog,
-      payload: {
-        id: deviceId
-      }
-    })
-    var mockPush = mocks.mockPush()
-    var accountRoutes = makeRoutes({
-      db: mockDB,
-      log: mockLog,
-      push: mockPush
-    })
-    var route = getRoute(accountRoutes, '/account/device/destroy')
-
-    return runTest(route, mockRequest, function () {
-      assert.equal(mockDB.deleteDevice.callCount, 1)
-
-      assert.equal(mockPush.notifyDeviceDisconnected.callCount, 1)
-      assert.equal(mockPush.notifyDeviceDisconnected.firstCall.args[0], mockRequest.auth.credentials.uid)
-      assert.equal(mockPush.notifyDeviceDisconnected.firstCall.args[1], deviceId)
-
-      assert.equal(mockLog.activityEvent.callCount, 1, 'log.activityEvent was called once')
-      var args = mockLog.activityEvent.args[0]
-      assert.equal(args.length, 1, 'log.activityEvent was passed one argument')
-      assert.deepEqual(args[0], {
-        event: 'device.deleted',
-        service: undefined,
-        userAgent: 'test user-agent',
-        uid: uid.toString('hex'),
-        device_id: deviceId
-      }, 'event data was correct')
-
-      assert.equal(mockLog.notifyAttachedServices.callCount, 1)
-      args = mockLog.notifyAttachedServices.args[0]
-      assert.equal(args.length, 3)
-      assert.equal(args[0], 'device:delete')
-      assert.equal(args[1], mockRequest)
-      var details = args[2]
-      assert.equal(details.uid, uid.toString('hex'))
-      assert.equal(details.id, deviceId)
-      assert.ok(Date.now() - details.timestamp < 100)
     })
   })
 })
@@ -1613,178 +1068,6 @@ describe('/account/login', function () {
   })
 })
 
-describe('/recovery_email/verify_code', function () {
-  var uid = uuid.v4('binary').toString('hex')
-  const mockLog = mocks.spyLog()
-  const mockRequest = mocks.mockRequest({
-    log: mockLog,
-    metricsContext: mocks.mockMetricsContext({
-      gather (data) {
-        return Promise.resolve(Object.assign(data, {
-          flowCompleteSignal: 'account.signed',
-          flow_time: 10000,
-          flow_id: 'F1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF103',
-          time: Date.now() - 10000
-        }))
-      }
-    }),
-    query: {},
-    payload: {
-      uid: uid,
-      code: 'e3c5b0e3f5391e134596c27519979b93',
-      service: 'sync'
-    }
-  })
-  var dbData = {
-    email: TEST_EMAIL,
-    emailCode: Buffer(mockRequest.payload.code, 'hex'),
-    emailVerified: false,
-    uid: uid
-  }
-  var dbErrors = {
-    verifyTokens: error.invalidVerificationCode({})
-  }
-  var mockDB = mocks.mockDB(dbData, dbErrors)
-  var mockMailer = mocks.mockMailer()
-  const mockPush = mocks.mockPush()
-  var mockCustoms = mocks.mockCustoms()
-  var accountRoutes = makeRoutes({
-    checkPassword: function () {
-      return P.resolve(true)
-    },
-    config: {},
-    customs: mockCustoms,
-    db: mockDB,
-    log: mockLog,
-    mailer: mockMailer,
-    push: mockPush
-  })
-  var route = getRoute(accountRoutes, '/recovery_email/verify_code')
-  describe('verifyTokens rejects with INVALID_VERIFICATION_CODE', function () {
-
-    it('without a reminder payload', function () {
-      return runTest(route, mockRequest, function (response) {
-        assert.equal(mockDB.verifyTokens.callCount, 1, 'calls verifyTokens')
-        assert.equal(mockDB.verifyEmail.callCount, 1, 'calls verifyEmail')
-        assert.equal(mockCustoms.check.callCount, 1, 'calls customs.check')
-        assert.equal(mockLog.notifyAttachedServices.callCount, 1, 'logs verified')
-
-        assert.equal(mockMailer.sendPostVerifyEmail.callCount, 1, 'sendPostVerifyEmail was called once')
-
-        assert.equal(mockLog.activityEvent.callCount, 1, 'activityEvent was called once')
-        let args = mockLog.activityEvent.args[0]
-        assert.equal(args.length, 1, 'log.activityEvent was passed one argument')
-        assert.deepEqual(args[0], {
-          event: 'account.verified',
-          service: 'sync',
-          userAgent: 'test user-agent',
-          uid: uid.toString('hex')
-        }, 'event data was correct')
-
-        assert.equal(mockLog.flowEvent.callCount, 2, 'flowEvent was called twice')
-        assert.equal(mockLog.flowEvent.args[0][0].event, 'email.verify_code.clicked', 'first event was email.verify_code.clicked')
-        assert.equal(mockLog.flowEvent.args[1][0].event, 'account.verified', 'second event was event account.verified')
-
-        assert.equal(mockPush.notifyUpdate.callCount, 1, 'mockPush.notifyUpdate should have been called once')
-        args = mockPush.notifyUpdate.args[0]
-        assert.equal(args.length, 2, 'mockPush.notifyUpdate should have been passed two arguments')
-        assert.equal(args[0].toString('hex'), uid, 'first argument should have been uid')
-        assert.equal(args[1], 'accountVerify', 'second argument should have been reason')
-
-        assert.equal(JSON.stringify(response), '{}')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockDB.verifyEmail.reset()
-        mockLog.activityEvent.reset()
-        mockLog.flowEvent.reset()
-        mockLog.notifyAttachedServices.reset()
-        mockMailer.sendPostVerifyEmail.reset()
-        mockPush.notifyUpdate.reset()
-      })
-    })
-
-    it('with a reminder payload', function () {
-      mockRequest.payload.reminder = 'second'
-
-      return runTest(route, mockRequest, function (response) {
-        assert.equal(mockLog.activityEvent.callCount, 1, 'activityEvent was called once')
-
-        assert.equal(mockLog.flowEvent.callCount, 3, 'flowEvent was called thrice')
-        assert.equal(mockLog.flowEvent.args[0][0].event, 'email.verify_code.clicked', 'first event was email.verify_code.clicked')
-        assert.equal(mockLog.flowEvent.args[1][0].event, 'account.verified', 'second event was account.verified')
-        assert.equal(mockLog.flowEvent.args[2][0].event, 'account.reminder', 'third event was account.reminder')
-
-        assert.equal(mockMailer.sendPostVerifyEmail.callCount, 1, 'sendPostVerifyEmail was called once')
-        assert.equal(mockPush.notifyUpdate.callCount, 1, 'mockPush.notifyUpdate should have been called once')
-
-        assert.equal(JSON.stringify(response), '{}')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockDB.verifyEmail.reset()
-        mockLog.activityEvent.reset()
-        mockLog.flowEvent.reset()
-        mockLog.notifyAttachedServices.reset()
-        mockMailer.sendPostVerifyEmail.reset()
-        mockPush.notifyUpdate.reset()
-      })
-    })
-  })
-
-  describe('verifyTokens resolves', function () {
-
-    before(() => {
-      dbData.emailVerified = true
-      dbErrors.verifyTokens = undefined
-    })
-
-    it('email verification', function () {
-      return runTest(route, mockRequest, function (response) {
-        assert.equal(mockDB.verifyTokens.callCount, 1, 'call db.verifyTokens')
-        assert.equal(mockDB.verifyEmail.callCount, 0, 'does not call db.verifyEmail')
-        assert.equal(mockLog.notifyAttachedServices.callCount, 0, 'does not call log.notifyAttachedServices')
-        assert.equal(mockLog.activityEvent.callCount, 0, 'log.activityEvent was not called')
-        assert.equal(mockPush.notifyUpdate.callCount, 0, 'mockPush.notifyUpdate should not have been called')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-      })
-    })
-
-    it('sign-in confirmation', function () {
-      dbData.emailCode = crypto.randomBytes(16)
-
-      return runTest(route, mockRequest, function (response) {
-        assert.equal(mockDB.verifyTokens.callCount, 1, 'call db.verifyTokens')
-        assert.equal(mockDB.verifyEmail.callCount, 0, 'does not call db.verifyEmail')
-        assert.equal(mockLog.notifyAttachedServices.callCount, 0, 'does not call log.notifyAttachedServices')
-
-        assert.equal(mockLog.activityEvent.callCount, 1, 'log.activityEvent was called once')
-        var args = mockLog.activityEvent.args[0]
-        assert.equal(args.length, 1, 'log.activityEvent was passed one argument')
-        assert.deepEqual(args[0], {
-          event: 'account.confirmed',
-          service: 'sync',
-          userAgent: 'test user-agent',
-          uid: uid.toString('hex')
-        }, 'event data was correct')
-
-        assert.equal(mockPush.notifyUpdate.callCount, 1, 'mockPush.notifyUpdate should have been called once')
-        args = mockPush.notifyUpdate.args[0]
-        assert.equal(args.length, 2, 'mockPush.notifyUpdate should have been passed two arguments')
-        assert.equal(args[0].toString('hex'), uid, 'first argument should have been uid')
-        assert.equal(args[1], 'accountConfirm', 'second argument should have been reason')
-      })
-      .then(function () {
-        mockDB.verifyTokens.reset()
-        mockLog.activityEvent.reset()
-        mockPush.notifyUpdate.reset()
-      })
-    })
-  })
-})
-
 describe('/account/keys', function () {
   var keyFetchTokenId = crypto.randomBytes(16)
   var uid = uuid.v4('binary')
@@ -1903,180 +1186,6 @@ describe('/account/destroy', function () {
         userAgent: 'test user-agent',
         uid: uid.toString('hex')
       }, 'event data was correct')
-    })
-  })
-})
-
-describe('/account/devices', function () {
-  it('should return the devices list', () => {
-    var mockRequest = mocks.mockRequest({
-      credentials: {
-        uid: crypto.randomBytes(16),
-        tokenId: crypto.randomBytes(16)
-      },
-      payload: {}
-    })
-    var unnamedDevice = { sessionToken: crypto.randomBytes(16) }
-    var mockDB = mocks.mockDB({
-      devices: [
-        { name: 'current session', type: 'mobile', sessionToken: mockRequest.auth.credentials.tokenId },
-        { name: 'has no type', sessionToken: crypto.randomBytes(16) },
-        { name: 'has device type', sessionToken: crypto.randomBytes(16), uaDeviceType: 'wibble' },
-        unnamedDevice
-      ]
-    })
-    var mockDevices = mocks.mockDevices()
-    var accountRoutes = makeRoutes({
-      db: mockDB,
-      devices: mockDevices
-    })
-    var route = getRoute(accountRoutes, '/account/devices')
-
-    return runTest(route, mockRequest, function (response) {
-      assert.ok(Array.isArray(response), 'response is array')
-      assert.equal(response.length, 4, 'response contains 4 items')
-
-      assert.equal(response[0].name, 'current session')
-      assert.equal(response[0].type, 'mobile')
-      assert.equal(response[0].sessionToken, undefined)
-      assert.equal(response[0].isCurrentDevice, true)
-
-      assert.equal(response[1].name, 'has no type')
-      assert.equal(response[1].type, 'desktop')
-      assert.equal(response[1].sessionToken, undefined)
-      assert.equal(response[1].isCurrentDevice, false)
-
-      assert.equal(response[2].name, 'has device type')
-      assert.equal(response[2].type, 'wibble')
-      assert.equal(response[2].isCurrentDevice, false)
-
-      assert.equal(response[3].name, null)
-
-      assert.equal(mockDB.devices.callCount, 1, 'db.devices was called once')
-      assert.equal(mockDB.devices.args[0].length, 1, 'db.devices was passed one argument')
-      assert.deepEqual(mockDB.devices.args[0][0], mockRequest.auth.credentials.uid, 'db.devices was passed uid')
-
-      assert.equal(mockDevices.synthesizeName.callCount, 1, 'mockDevices.synthesizeName was called once')
-      assert.equal(mockDevices.synthesizeName.args[0].length, 1, 'mockDevices.synthesizeName was passed one argument')
-      assert.equal(mockDevices.synthesizeName.args[0][0], unnamedDevice, 'mockDevices.synthesizeName was passed unnamed device')
-    })
-  })
-})
-
-describe('/account/login/send_unblock_code', function () {
-  var uid = uuid.v4('binary').toString('hex')
-  var email = 'unblock@example.com'
-  const mockLog = mocks.spyLog()
-  var mockRequest = mocks.mockRequest({
-    log: mockLog,
-    payload: {
-      email: email,
-      metricsContext: {
-        flowBeginTime: Date.now(),
-        flowId: 'F1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF1031DF103'
-      }
-    }
-  })
-  var mockMailer = mocks.mockMailer()
-  var mockDb = mocks.mockDB({
-    uid: uid,
-    email: email
-  })
-  var config = {
-    signinUnblock: {
-      allowedEmailAddresses: /unblock.*$/
-    }
-  }
-  var accountRoutes = makeRoutes({
-    config: config,
-    db: mockDb,
-    log: mockLog,
-    mailer: mockMailer
-  })
-  var route = getRoute(accountRoutes, '/account/login/send_unblock_code')
-
-  afterEach(function () {
-    mockDb.emailRecord.reset()
-    mockDb.createUnblockCode.reset()
-    mockMailer.sendUnblockCode.reset()
-  })
-
-  it('signin unblock enabled', function () {
-    config.signinUnblock.enabled = true
-    return runTest(route, mockRequest, function (response) {
-      assert.ok(!(response instanceof Error), response.stack)
-      assert.deepEqual(response, {}, 'response has no keys')
-
-      assert.equal(mockDb.emailRecord.callCount, 1, 'db.emailRecord called')
-      assert.equal(mockDb.emailRecord.args[0][0], email)
-
-      assert.equal(mockDb.createUnblockCode.callCount, 1, 'db.createUnblockCode called')
-      var dbArgs = mockDb.createUnblockCode.args[0]
-      assert.equal(dbArgs.length, 1)
-      assert.equal(dbArgs[0], uid)
-
-      assert.equal(mockMailer.sendUnblockCode.callCount, 1, 'called mailer.sendUnblockCode')
-      var args = mockMailer.sendUnblockCode.args[0]
-      assert.equal(args.length, 3, 'mailer.sendUnblockCode called with 3 args')
-
-      assert.equal(mockLog.flowEvent.callCount, 1, 'log.flowEvent was called once')
-      assert.equal(mockLog.flowEvent.args[0][0].event, 'account.login.sentUnblockCode', 'event was account.login.sentUnblockCode')
-      mockLog.flowEvent.reset()
-    })
-  })
-
-  it('signin unblock disabled', function () {
-    config.signinUnblock.enabled = false
-
-    return runTest(route, mockRequest).then(() => assert.ok(false), err => {
-      assert.equal(err.output.statusCode, 503, 'correct status code is returned')
-      assert.equal(err.errno, error.ERRNO.FEATURE_NOT_ENABLED, 'correct errno is returned')
-
-      assert.equal(mockLog.flowEvent.callCount, 0, 'log.flowEvent was not called')
-    })
-  })
-
-  it('uses normalized email address for feature flag', function () {
-    config.signinUnblock.enabled = true
-    mockRequest.payload.email = 'UNBLOCK@example.com'
-
-    return runTest(route, mockRequest, function(response) {
-      assert.ok(!(response instanceof Error), response.stack)
-      assert.deepEqual(response, {}, 'response has no keys')
-
-      assert.equal(mockDb.emailRecord.callCount, 1, 'db.emailRecord called')
-      assert.equal(mockDb.emailRecord.args[0][0], mockRequest.payload.email)
-      assert.equal(mockDb.createUnblockCode.callCount, 1, 'db.createUnblockCode called')
-      assert.equal(mockMailer.sendUnblockCode.callCount, 1, 'called mailer.sendUnblockCode')
-    })
-  })
-})
-
-describe('/account/login/reject_unblock_code', function () {
-  it('should consume the unblock code', () => {
-    var uid = uuid.v4('binary').toString('hex')
-    var unblockCode = 'A1B2C3D4'
-    var mockRequest = mocks.mockRequest({
-      payload: {
-        uid: uid,
-        unblockCode: unblockCode
-      }
-    })
-    var mockDb = mocks.mockDB()
-    var accountRoutes = makeRoutes({
-      db: mockDb
-    })
-    var route = getRoute(accountRoutes, '/account/login/reject_unblock_code')
-
-    return runTest(route, mockRequest, function (response) {
-      assert.ok(!(response instanceof Error), response.stack)
-      assert.deepEqual(response, {}, 'response has no keys')
-
-      assert.equal(mockDb.consumeUnblockCode.callCount, 1, 'consumeUnblockCode is called')
-      var args = mockDb.consumeUnblockCode.args[0]
-      assert.equal(args.length, 2)
-      assert.equal(args[0].toString('hex'), uid)
-      assert.equal(args[1], unblockCode)
     })
   })
 })
